libs: Fix FD comparisons in binder
Even though 0 is a valid file descriptor (FD),
all checks made using it as error.

When a user space process is started by the kernel,
there are no FD open.

The first one to be opened is 0. If this process used binder,
then its connection would be dropped due to the checks mentioned
previously.

This issue is fixed by changing the comparisons to allow a FD 0.

This change was created by Alexandre while he was an employee of
Samsung Electronics (SRBR). He is no longer part of the staff.
On behalf of Samsung Electronics (SRBR).

BUG: 122699850
Test: Called test binary from kernel and ensured connection
to libbinder is not dropped.

Change-Id: I588ec8c4c1ba130ea2d5d61d94dea66838e4aa1d
Signed-off-by: ghou8s <ghou8s@gmail.com>